Use read bytes instead of serialising struct
The network config JSON may be passed on by the main plugin to the
IPAM plugin. If we serialise a plugin.NetConf struct though, it will
only include the fields in plugin.NetConf -- so, not anything in
IPAM. The result is that the IPAM plugin doesn't get any
configuration, and returns an error.