Move argument-handling and conf-loading into public API
The fact that the arguments get turned into an environment for the
plugin is incidental to the API; so, provide a way of supplying them
as a struct or saying "just use the same arguments as I got" (the
latter is for IPAM plugins).

It's also worth having some mix-and-match procedures for finding and
loading configuration from the filesystem, since applications will
want to have different conventions. (These could probably go in their
own pkg/.)